+++ CHANGES.rst
@@ -10,3 +10,5 @@ Changelog for cookiecutter-djangosite-template
 - Got first version of the template to work.
 
 - Added docker test setup, Jenkins file and ansible integration.
+
+- Fixed template configuration for recent django versions.